+++ stroom-annotation/stroom-annotation-impl/src/main/java/stroom/annotation/impl/AnnotationConfig.java
@@ -30,16 +30,22 @@ import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import jakarta.validation.constraints.Min;
+
+import java.util.Objects;
 
 @JsonPropertyOrder(alphabetic = true)
 public class AnnotationConfig extends AbstractConfig implements IsStroomConfig, HasDbConfig {
 
     public static final String DEFAULT_RETENTION_PERIOD = "5y";
+    public static final int DEFAULT_EVENT_LINK_CACHE_SIZE_LIMIT = 1_000_000;
+    public static final String EVENT_LINK_CACHE_SIZE_LIMIT_PROP_NAME = "eventLinkCacheSizeLimit";
 
     private final AnnotationDBConfig dbConfig;
     private final String createText;
     private final String defaultRetentionPeriod;
     private final StroomDuration physicalDeleteAge;
+    private final int eventLinkCacheSizeLimit;
     private final CacheConfig annotationTagCache;
     private final CacheConfig annotationFeedCache;
     private final CacheConfig annotationValCache;
@@ -50,6 +56,7 @@ public class AnnotationConfig extends AbstractConfig implements IsStroomConfig, 
         createText = "Create Annotation";
         defaultRetentionPeriod = DEFAULT_RETENTION_PERIOD;
         physicalDeleteAge = StroomDuration.ofDays(7);
+        eventLinkCacheSizeLimit = DEFAULT_EVENT_LINK_CACHE_SIZE_LIMIT;
         annotationTagCache = CacheConfig.builder()
                 .maximumSize(1000L)
                 .expireAfterAccess(StroomDuration.ofMinutes(10))
@@ -76,6 +83,7 @@ public class AnnotationConfig extends AbstractConfig implements IsStroomConfig, 
                             @JsonProperty("createText") final String createText,
                             @JsonProperty("defaultRetentionPeriod") final String defaultRetentionPeriod,
                             @JsonProperty("physicalDeleteAge") final StroomDuration physicalDeleteAge,
+                            @JsonProperty(EVENT_LINK_CACHE_SIZE_LIMIT_PROP_NAME) final Integer eventLinkCacheSizeLimit,
                             @JsonProperty("annotationTagCache") final CacheConfig annotationTagCache,
                             @JsonProperty("annotationFeedCache") final CacheConfig annotationFeedCache,
                             @JsonProperty("annotationValCache") final CacheConfig annotationValCache,
@@ -84,6 +92,8 @@ public class AnnotationConfig extends AbstractConfig implements IsStroomConfig, 
         this.createText = createText;
         this.defaultRetentionPeriod = defaultRetentionPeriod;
         this.physicalDeleteAge = physicalDeleteAge;
+        this.eventLinkCacheSizeLimit = Objects.requireNonNullElse(
+                eventLinkCacheSizeLimit, DEFAULT_EVENT_LINK_CACHE_SIZE_LIMIT);
         this.annotationTagCache = annotationTagCache;
         this.annotationFeedCache = annotationFeedCache;
         this.annotationValCache = annotationValCache;
@@ -112,6 +122,13 @@ public class AnnotationConfig extends AbstractConfig implements IsStroomConfig, 
                              "In ISO-8601 duration format, e.g. 'P1DT12H'")
     public StroomDuration getPhysicalDeleteAge() {
         return physicalDeleteAge;
+    }
+
+    @JsonPropertyDescription("The maximum number of annotation to event links that can be cached. If this " +
+                             "number is breached, it will result in an error.")
+    @Min(0)
+    public int getEventLinkCacheSizeLimit() {
+        return eventLinkCacheSizeLimit;
     }
 
     @JsonPropertyDescription("Cache config for annotation tags")